+++ fabric/src/main/java/com/avp/fabric/data/advancement/AdvancementProvider.java
@@ -1,5 +1,7 @@
 package com.avp.fabric.data.advancement;
 
+import com.alien.common.data.AlienAdvancements;
+import com.alien.common.registry.init.AlienBlocks;
 import com.alien.common.registry.init.AlienEntityTypes;
 import com.alien.common.registry.init.AlienItems;
 import com.alien.common.registry.init.block.AlienResinBlocks;
@@ -11,14 +13,15 @@ import net.minecraft.advancements.AdvancementHolder;
 import net.minecraft.advancements.AdvancementRequirements;
 import net.minecraft.advancements.AdvancementRewards;
 import net.minecraft.advancements.AdvancementType;
+import net.minecraft.advancements.CriteriaTriggers;
 import net.minecraft.advancements.critereon.EntityPredicate;
+import net.minecraft.advancements.critereon.ImpossibleTrigger;
 import net.minecraft.advancements.critereon.InventoryChangeTrigger;
 import net.minecraft.advancements.critereon.ItemPredicate;
 import net.minecraft.advancements.critereon.KilledTrigger;
 import net.minecraft.advancements.critereon.PlayerInteractTrigger;
 import net.minecraft.core.HolderLookup;
 import net.minecraft.core.registries.BuiltInRegistries;
-import net.minecraft.network.chat.Component;
 import net.minecraft.tags.TagKey;
 import net.minecraft.world.entity.EntityType;
 import net.minecraft.world.item.Items;
@@ -29,7 +32,6 @@ import java.util.Optional;
 import java.util.concurrent.CompletableFuture;
 import java.util.function.Consumer;
 
-import com.avp.AVP;
 import com.avp.AVPResources;
 import com.avp.common.registry.tag.AVPEntityTypeTags;
 
@@ -110,8 +112,8 @@ public class AdvancementProvider extends FabricAdvancementProvider {
         var root = Advancement.Builder.advancement()
             .display(
                 AlienResinBlocks.RESIN.get(),
-                Component.translatable("advancements.aliens.root.title"),
-                Component.translatable("advancements.aliens.root.description"),
+                AlienAdvancements.ROOT.getTitleComponent(),
+                AlienAdvancements.ROOT.getDescriptionComponent(),
                 AVPResources.location("textures/gui/advancements/backgrounds/resin.png"),
                 AdvancementType.TASK,
                 false,
@@ -119,18 +121,21 @@ public class AdvancementProvider extends FabricAdvancementProvider {
                 false
             )
             .addCriterion("crafting_table", InventoryChangeTrigger.TriggerInstance.hasItems(Blocks.CRAFTING_TABLE))
-            .save(consumer, AVP.MOD_ID + ":aliens/root");
+            .save(consumer, AlienAdvancements.ROOT.getResourceLocation().toString());
+
+        var removeEmbryoWithChorusFruitAdvancement = addRemoveEmbryoWithChorusFruitAdvancement(root, consumer);
 
         var alienKillerAdvancement = addAlienKillerAdvancement(root, consumer);
         var royalAlienKillerAdvancement = addRoyalAlienKillerAdvancement(alienKillerAdvancement, consumer);
+        var hiveBusterAdvancement = addHiveBusterAdvancement(royalAlienKillerAdvancement, consumer);
         var xenocideAdvancement = addXenocideAdvancement(royalAlienKillerAdvancement, consumer);
 
         var shearAnOvomorphAdvancement = Advancement.Builder.advancement()
-            .parent(root)
+            .parent(alienKillerAdvancement)
             .display(
                 Items.SHEARS,
-                Component.translatable("advancements.aliens.shear_an_ovomorph.title"),
-                Component.translatable("advancements.aliens.shear_an_ovomorph.description"),
+                AlienAdvancements.SHEAR_AN_OVOMORPH.getTitleComponent(),
+                AlienAdvancements.SHEAR_AN_OVOMORPH.getDescriptionComponent(),
                 null,
                 AdvancementType.TASK,
                 true,
@@ -138,13 +143,13 @@ public class AdvancementProvider extends FabricAdvancementProvider {
                 false
             )
             .addCriterion(
-                "shear_an_ovomorph",
+                AlienAdvancements.SHEAR_AN_OVOMORPH.path(),
                 PlayerInteractTrigger.TriggerInstance.itemUsedOnEntity(
                     ItemPredicate.Builder.item().of(Items.SHEARS),
                     Optional.of(EntityPredicate.wrap(EntityPredicate.Builder.entity().of(AVPEntityTypeTags.OVOMORPHS)))
                 )
             )
-            .save(consumer, AVP.MOD_ID + ":aliens/shear_an_ovomorph");
+            .save(consumer, AlienAdvancements.SHEAR_AN_OVOMORPH.getResourceLocation().toString());
 
         var addChitinArmorAdvancement = addChitinArmorAdvancements(alienKillerAdvancement, consumer);
 
@@ -156,13 +161,22 @@ public class AdvancementProvider extends FabricAdvancementProvider {
             .parent(parent)
             .display(
                 AlienArmorItems.CHITIN_HELMET.get(),
-                Component.translatable("advancements.aliens.chitin_armor.title"),
-                Component.translatable("advancements.aliens.chitin_armor.description"),
+                AlienAdvancements.WEAR_CHITIN_ARMOR.getTitleComponent(),
+                AlienAdvancements.WEAR_CHITIN_ARMOR.getDescriptionComponent(),
                 null,
                 AdvancementType.TASK,
                 true,
                 true,
                 false
+            )
+            .addCriterion(
+                "aberrant_chitin_armor",
+                InventoryChangeTrigger.TriggerInstance.hasItems(
+                    AlienArmorItems.ABERRANT_CHITIN_HELMET.get(),
+                    AlienArmorItems.ABERRANT_CHITIN_CHESTPLATE.get(),
+                    AlienArmorItems.ABERRANT_CHITIN_LEGGINGS.get(),
+                    AlienArmorItems.ABERRANT_CHITIN_BOOTS.get()
+                )
             )
             .addCriterion(
                 "chitin_armor",
@@ -183,7 +197,7 @@ public class AdvancementProvider extends FabricAdvancementProvider {
                 )
             )
             .requirements(AdvancementRequirements.Strategy.OR)
-            .save(consumer, AVP.MOD_ID + ":aliens/chitin_armor");
+            .save(consumer, AlienAdvancements.WEAR_CHITIN_ARMOR.getResourceLocation().toString());
     }
 
     private AdvancementHolder addPlatedChitinArmorAdvancement(AdvancementHolder parent, Consumer<AdvancementHolder> consumer) {
@@ -191,13 +205,22 @@ public class AdvancementProvider extends FabricAdvancementProvider {
             .parent(parent)
             .display(
                 AlienArmorItems.PLATED_CHITIN_HELMET.get(),
-                Component.translatable("advancements.aliens.plated_chitin_armor.title"),
-                Component.translatable("advancements.aliens.plated_chitin_armor.description"),
+                AlienAdvancements.WEAR_PLATED_CHITIN_ARMOR.getTitleComponent(),
+                AlienAdvancements.WEAR_PLATED_CHITIN_ARMOR.getDescriptionComponent(),
                 null,
                 AdvancementType.CHALLENGE,
                 true,
                 true,
                 false
+            )
+            .addCriterion(
+                "plated_aberrant_chitin_armor",
+                InventoryChangeTrigger.TriggerInstance.hasItems(
+                    AlienArmorItems.PLATED_ABERRANT_CHITIN_HELMET.get(),
+                    AlienArmorItems.PLATED_ABERRANT_CHITIN_CHESTPLATE.get(),
+                    AlienArmorItems.PLATED_ABERRANT_CHITIN_LEGGINGS.get(),
+                    AlienArmorItems.PLATED_ABERRANT_CHITIN_BOOTS.get()
+                )
             )
             .addCriterion(
                 "plated_chitin_armor",
@@ -219,7 +242,7 @@ public class AdvancementProvider extends FabricAdvancementProvider {
             )
             .requirements(AdvancementRequirements.Strategy.OR)
             .rewards(AdvancementRewards.Builder.experience(100))
-            .save(consumer, AVP.MOD_ID + ":aliens/plated_chitin_armor");
+            .save(consumer, AlienAdvancements.WEAR_PLATED_CHITIN_ARMOR.getResourceLocation().toString());
     }
 
     private AdvancementHolder addAlienKillerAdvancement(AdvancementHolder parent, Consumer<AdvancementHolder> consumer) {
@@ -227,8 +250,8 @@ public class AdvancementProvider extends FabricAdvancementProvider {
             .parent(parent)
             .display(
                 AlienItems.CHITIN.get(),
-                Component.translatable("advancements.aliens.kill_an_alien.title"),
-                Component.translatable("advancements.aliens.kill_an_alien.description"),
+                AlienAdvancements.KILL_AN_ALIEN.getTitleComponent(),
+                AlienAdvancements.KILL_AN_ALIEN.getDescriptionComponent(),
                 null,
                 AdvancementType.TASK,
                 true,
@@ -236,7 +259,7 @@ public class AdvancementProvider extends FabricAdvancementProvider {
                 false
             )
             .requirements(AdvancementRequirements.Strategy.OR)
-            .save(consumer, AVP.MOD_ID + ":aliens/kill_an_alien");
+            .save(consumer, AlienAdvancements.KILL_AN_ALIEN.getResourceLocation().toString());
     }
 
     private AdvancementHolder addRoyalAlienKillerAdvancement(AdvancementHolder parent, Consumer<AdvancementHolder> consumer) {
@@ -244,8 +267,8 @@ public class AdvancementProvider extends FabricAdvancementProvider {
             .parent(parent)
             .display(
                 AlienItems.PLATED_CHITIN.get(),
-                Component.translatable("advancements.aliens.kill_a_royal_alien.title"),
-                Component.translatable("advancements.aliens.kill_a_royal_alien.description"),
+                AlienAdvancements.KILL_A_ROYAL_ALIEN.getTitleComponent(),
+                AlienAdvancements.KILL_A_ROYAL_ALIEN.getDescriptionComponent(),
                 null,
                 AdvancementType.TASK,
                 true,
@@ -253,16 +276,16 @@ public class AdvancementProvider extends FabricAdvancementProvider {
                 false
             )
             .requirements(AdvancementRequirements.Strategy.OR)
-            .save(consumer, AVP.MOD_ID + ":aliens/kill_a_royal_alien");
+            .save(consumer, AlienAdvancements.KILL_A_ROYAL_ALIEN.getResourceLocation().toString());
     }
 
     private AdvancementHolder addXenocideAdvancement(AdvancementHolder parent, Consumer<AdvancementHolder> consumer) {
         return addMobsToKill(Advancement.Builder.advancement(), ALIENS_TO_KILL)
             .parent(parent)
             .display(
-                AlienItems.RAW_ROYAL_JELLY.get(),
-                Component.translatable("advancements.aliens.kill_all_aliens.title"),
-                Component.translatable("advancements.aliens.kill_all_aliens.description"),
+                AlienBlocks.ROYAL_JELLY_BLOCK.get(),
+                AlienAdvancements.KILL_ALL_ALIENS.getTitleComponent(),
+                AlienAdvancements.KILL_ALL_ALIENS.getDescriptionComponent(),
                 null,
                 AdvancementType.CHALLENGE,
                 true,
@@ -271,7 +294,45 @@ public class AdvancementProvider extends FabricAdvancementProvider {
             )
             .requirements(AdvancementRequirements.Strategy.AND)
             .rewards(AdvancementRewards.Builder.experience(100))
-            .save(consumer, AVP.MOD_ID + ":aliens/kill_all_aliens");
+            .save(consumer, AlienAdvancements.KILL_ALL_ALIENS.getResourceLocation().toString());
+    }
+
+    private AdvancementHolder addHiveBusterAdvancement(AdvancementHolder parent, Consumer<AdvancementHolder> consumer) {
+        return Advancement.Builder.advancement()
+            .addCriterion("kill_a_hive", CriteriaTriggers.IMPOSSIBLE.createCriterion(new ImpossibleTrigger.TriggerInstance()))
+            .parent(parent)
+            .display(
+                AlienItems.RAW_ROYAL_JELLY.get(),
+                AlienAdvancements.KILL_A_HIVE.getTitleComponent(),
+                AlienAdvancements.KILL_A_HIVE.getDescriptionComponent(),
+                null,
+                AdvancementType.CHALLENGE,
+                true,
+                true,
+                false
+            )
+            .rewards(AdvancementRewards.Builder.experience(100))
+            .save(consumer, AlienAdvancements.KILL_A_HIVE.getResourceLocation().toString());
+    }
+
+    private AdvancementHolder addRemoveEmbryoWithChorusFruitAdvancement(AdvancementHolder parent, Consumer<AdvancementHolder> consumer) {
+        return Advancement.Builder.advancement()
+            .addCriterion(
+                "remove_embryo_with_chorus_fruit",
+                CriteriaTriggers.IMPOSSIBLE.createCriterion(new ImpossibleTrigger.TriggerInstance())
+            )
+            .parent(parent)
+            .display(
+                Items.CHORUS_FRUIT,
+                AlienAdvancements.REMOVE_EMBRYO_WITH_CHORUS_FRUIT.getTitleComponent(),
+                AlienAdvancements.REMOVE_EMBRYO_WITH_CHORUS_FRUIT.getDescriptionComponent(),
+                null,
+                AdvancementType.TASK,
+                true,
+                true,
+                false
+            )
+            .save(consumer, AlienAdvancements.REMOVE_EMBRYO_WITH_CHORUS_FRUIT.getResourceLocation().toString());
     }
 
     private Advancement.Builder addMobsToKill(Advancement.Builder builder, String criterionKey, TagKey<EntityType<?>> entityTypeTagKey) {